Add files via upload
Support Vector Machine is considered to be one of the black box techniques, because in this we cannot determine what are the actual calculations and operations are being performed on the data to get the outcome that it shows. So working on it is pretty straight forward since we don't need to do much tweaking and tuning of the data manually.